Create Temperature converter website
Temperature Converter Website The user will input a temperature in either Fahrenheit or Celsius or Kelvin and press a convert button. The converted temperature will then be displayed with the correct unit. Your UI should include four main elements. First, input field for the user to type in the temperature-don't forget to validate that the input is a number!
 Second, drop-down menu or radio button for the user to choose if they are inputting in Celsius or Fahrenheit.
 Third, convert button to trigger the temperature conversion.
 Fourth, display area for the finally converted temperature and unit.