Create A basic To-Do webapps
Develop a basic to-do app that enables users to add daily tasks. The to-do web app must allow users to add new items to a list, and items added to a list must be visible. Also, a user can mark a task complete upon its successful completion, and this completed task must appear on a separated list se completed task. You can also add a separate section titled Pending Task that will list all the tasks that acquired completion. In such a case, the to-do list will be a union of a pending task list and the completed task list. irrespective of a list, A user must be able to delete or edit tasks from it. To further increase the web app's complexity or make it more detailed, you can add the date and time when a task is added and/or completed.
Thank You #futureIntern for giving me this great opportunity